refactor: DragHandle 공통 컴포넌트로 추출
