+++ src/app/(main)/template/_components/CommonContentTable/CommonContentTable.css.ts
@@ -59,34 +59,6 @@ export const tdStyle = style({
   },
 })
 
-export const dragHandleStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '3px',
-  cursor: 'grab',
-  color: colors.gray300,
-  flexShrink: 0,
-  width: '16px',
-  alignItems: 'center',
-  selectors: {
-    '&:active': {
-      cursor: 'grabbing',
-    },
-  },
-})
-
-export const dragDotRowStyle = style({
-  display: 'flex',
-  gap: '2px',
-})
-
-export const dragDotStyle = style({
-  width: '3px',
-  height: '3px',
-  borderRadius: '50%',
-  backgroundColor: 'currentColor',
-})
-
 export const rowLabelStyle = style({
   marginLeft: '8px',
   fontSize: fontStyles.titleSm.fontSize,
