+++ src/app/(main)/template/_components/IndividualContentTable/IndividualContentTable.css.ts
@@ -173,30 +173,6 @@ export const cellButtonRecipe = recipe({
   variants: {},
 })
 
-export const dragHandleStyle = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '2px',
-  cursor: 'grab',
-  color: colors.gray300,
-  flexShrink: 0,
-  selectors: {
-    '&:active': { cursor: 'grabbing' },
-  },
-})
-
-export const dragDotRowStyle = style({
-  display: 'flex',
-  gap: '2px',
-})
-
-export const dragDotStyle = style({
-  width: '3px',
-  height: '3px',
-  borderRadius: '50%',
-  backgroundColor: 'currentColor',
-})
-
 export const dragOverlayWrapperStyle = style({
   display: 'flex',
   flexDirection: 'column',
