+++ src/components/common/DragHandle/DragHandle.tsx
@@ -1,0 +1,20 @@
+import { dragHandleStyle, dragDotRowStyle, dragDotStyle } from './DragHandle.css'
+
+export default function DragHandle() {
+  return (
+    <span className={dragHandleStyle}>
+      <span className={dragDotRowStyle}>
+        <span className={dragDotStyle} />
+        <span className={dragDotStyle} />
+      </span>
+      <span className={dragDotRowStyle}>
+        <span className={dragDotStyle} />
+        <span className={dragDotStyle} />
+      </span>
+      <span className={dragDotRowStyle}>
+        <span className={dragDotStyle} />
+        <span className={dragDotStyle} />
+      </span>
+    </span>
+  )
+}
